requirements.txt
streamlit
openai
requests
python-docx
python-pptx
Pillow
pandas
ollama
transformers
torch
accelerate
huggingface_hub
duckduckgo-search
langchain
langchain-community
python-dotenv